requirements.txt
# Core dependencies
numpy==1.24.3
pandas==2.0.2

# Deep Learning (CPU version for local inference)
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2

# Image processing
opencv-python==4.8.0.74
Pillow==10.0.0
scikit-image==0.21.0
scipy==1.11.1

# Geospatial
rasterio==1.3.8
geopandas==0.13.2
shapely==2.0.1
fiona==1.9.4
pyproj==3.6.0

# Data augmentation
albumentations==1.3.1

# Segmentation models
segmentation-models-pytorch==0.3.3
timm==0.9.5

# Machine learning utilities
scikit-learn==1.3.0

# Progress bars and utilities
tqdm==4.65.0
pyyaml==6.0.1

# Visualization
matplotlib==3.7.2
seaborn==0.12.2

# CLI utilities
click==8.1.7
